Add date and navigation modules
