+++ index.html
@@ -3,9 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script defer src="index.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/luxon@2.0.2/build/global/luxon.min.js" crossorigin="anonymous"></script>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />  
     <link rel="stylesheet" href="index.css" />
     <title>Awesome Books</title>
   </head>
@@ -60,6 +58,6 @@
       <footer>
           <p>&copy; Copyright 2022 Gegardus</p>
       </footer>  
-     
+      <script type="module" src="index.js"></script>        
   </body>  
   </html>
