+++ package.json
@@ -16,5 +16,8 @@
   "bugs": {
     "url": "https://github.com/Gegardus/Awesome-Books-With-ES6/issues"
   },
-  "homepage": "https://github.com/Gegardus/Awesome-Books-With-ES6#readme"
+  "homepage": "https://github.com/Gegardus/Awesome-Books-With-ES6#readme",
+  "dependencies": {
+    "luxon": "^2.3.0"
+  }
 }
